fix: rename files and move puppeteer files into a subdir
